// src/pages/systemConfig/consume/suit/data.d.ts
export type SuitType = {
  id?: string|number;
  title?: string;
  sort?: string
  yearFee?: number;
  refreshNum: number|string;
  communicateNum: number|string;
  resumeNum: number|string;
  videoNum: number|string;
  jobNum: number|string;
  memberNum: number|string;
  topNum: number|string;
  publishUser?: string;
  publishTime?: string;
};